modeling/requirements.txt
cuda-python==11.8.3
cudf-cu11==24.8.2
cupy-cuda11x==13.3.0
loguru==0.7.2
lxml==5.3.0
matplotlib==3.9.2
minio==7.2.8
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvtx==0.2.10
pandas==2.2.2
peft==0.13.0
pymilvus==2.4.5
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
tqdm==4.66.5
transformers==4.44.1
unlimited-machine-translator==1.1.8